add I2C6 peripherals for testing
Adds a device tree overlay adding an LTC4151 and a MAX31790 connected to
I2C6. This is the setup I am testing with at the moment. The main
purpose of this is to test that I can write device trees and have
buildroot build them for us, then script their installation into the
system image.